Prefer C++ method std::fmod over C method fmod

+++ common/include/pcl/common/impl/angles.hpp
@@ -48,11 +48,11 @@ namespace pcl
   normAngle (float alpha)
   {
     return (alpha >= 0  ? 
-        fmodf (alpha + static_cast<float>(M_PI), 
+        std::fmod (alpha + static_cast<float>(M_PI), 
                2.0f * static_cast<float>(M_PI)) 
         - static_cast<float>(M_PI) 
         : 
-        -(fmodf (static_cast<float>(M_PI) - alpha, 
+        -(std::fmod (static_cast<float>(M_PI) - alpha, 
                  2.0f * static_cast<float>(M_PI)) 
         - static_cast<float>(M_PI)));
   }
